+++ src/components/Admin/ManageProducts.js
@@ -13,7 +13,7 @@ import Button from '@material-ui/core/Button';
 import IconButton from '@material-ui/core/IconButton';
 import DeleteIcon from '@material-ui/icons/Delete';
 import EditIcon from '@material-ui/icons/Edit';
-import { deleteProduct } from '../reducers';
+import { deleteProduct } from '../../reducers';
 
 const styles = theme => ({
   root: {
